MODORDERS-328 # package renaming
